backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
openpyxl
mysql-connector-python
python-multipart
